basic: use "double" numbers
Performance is roughly on par with float, so apart from higher memory
usage, there is no reason not to use doubles.

+++ ttbasic/basic.cpp
@@ -49,7 +49,7 @@
 
 #define FLOAT_NUMS
 #ifdef FLOAT_NUMS
-typedef float num_t;
+typedef double num_t;
 #else
 typedef int32_t num_t;
 #endif
